Job requirement
Fixed job requirement and names according to english text from the site,
it seems "yen" and "circle" refers to same thing, and we currently don't
know what's the term of the requirement, is it "grade"? or "level"? but
these texts seems to be pointing to a requirement for skill or advancing
to next class

ETC_20150312_001869 says "Mugosa" but I don't know what class is "무고사"
exactly.