Update QUEST_LV_0100_en.tsv
First time contributing to the project. I'm still very new to this github thing ;)
I did some typos fixes, mainly ".." lines using the search function (Ctrl+f).
There was a line where I didn't change the ".." since it was machine translated and I wanted to give people an opportunity to translate it again in the future.
If i made any mistakes feel free to change it as well!

P.S.: There was also an weird thing on some lines starting with ")}", I didn't removed it since I thought it was there for some reason.