Update ITEM_en.tsv
String updates for consumables.

Lines 1819 and 1794 are the same name for two different items. 1819 (ITEM_20150317_001922) is an HP Increase Potion that increases your Health STAT while 1794 (ITEM_20150317_001897) is an HP Increase Potion that increase your MAX Health. Didn't know what to change so I just left it as-is, thought it would be useful to point this out.

Lines 1846 and 1847 are exactly the same Item name (Condensed HP Potion). Item numbers ITEM_20150317_001949 and ITEM_20150317_001950.

String updates for ETC items.

Lines 1871 and 1872 are also the exact same name (Gorgon Card). Item numbers ITEM_20150317_001974 and ITEM_20150317_001975.

Possible issue with line 2148 and 2149, I believe 2148 should be "Nudock Hook" and 2149 should end with "Nudocks".

Violet Flower is repeated in lines 2285 and 2037.

Maize Spore Pocket is repeated in lines 2314 and 2043.