Update ITEM_en.tsv
Editing for Spears

I didn't know if the " - " was intended at the start of the Trident line [Line 671] (if it were part of a script or something), but I deleted it guessing it was just a typo.

Line 664 is said to be a quote, so I put quotations.  However, I'm not sure if it's standard in the script for quotes to be true quotation marks or apostrophes, as some games use.