Update ETC.tsv
I don't speak Korean, but managed to deconstruction some sentences and use the wiki + entries on the TOSbase translate tool to translate a few of them (with not a little help from CTRL+f).  

Line 11917 - Ratas Gallery Site (라타스 회랑 터) the only one I'm not confident in.  I know the area is Ratas, but not sure what the 터 at the end denotes. 

Also not sure about Pond/Pondell as they're spelled Pawndell/pawnd in one one translation and Pond in another.  Assuming this is the correct spelling, at least for now.