Update QUEST_LV_0100.tsv
- Removed many unnecessary whitespaces
- Updated dialogue in lines 001 - 510
- Replaced Miner's Village with Miners' Village (unless there really
only is one miner and it's his village)

QUEST_LV_0100_20150317_000113 - previous line mentions assistant named
Irene and then uses "he" to refer to said assistant. Irene being a
womans name I changed this to she instead. Irene is also later referred
to as "her". but without context it's hard to know if this line is
actually referring to Irene at all. could be someone else entirely.
QUEST_LV_0100_20150317_000173 - changed Greak to Great as Greak isn't a
word and it doesn't seem to appear anywhere else in the document, so I'm
assuming it's a typo
QUEST_LV_0100_20150317_000255 - not sure what to make of this line
really
QUEST_LV_0100_20150317_000278 - I am not sure what to make of this bit
{nl}{-scp
SCR_DIALOG_NPC_ANIM( .. does it need to be there or not?
QUEST_LV_0100_20150317_000279 - a number of lines including this one
start with the symbols )}.. is there a reason for this?