Update QUEST.tsv
Reverted 1505 to the original since the intent is vague. The person might be saying he's not sure about the rebellion or he could be saying he's not sure about something that's not rebellion yet and then tries to shift the subject to the rebellion.

Changed 1596 as suggested. While we don't know there is one more monster or more than one more monster, "More monsters!" would apply to both situations unlike "Another monster!"

Reverted 1577 and 1746. Need to know the situation first, as vtange pointed out.

Thanks.