Update QUEST.tsv
Grammar checks, rewrites, etc. A lot of lines repeat in this section so I've made changes to each of them based on the new information that each had though some I left alone because of lack of context.

The name "Ebonipon" showed up but I can't be sure if it's the same as "Ebonippon" because the instances Ebonipon show up indicates it's a poison or group of people and the instances Ebonippon show up refer to an assassin named Ebonippon. I've ignored them for now.

Need to know if Gatus and Gitis at lines 1850-1851 are the same person.