Update QUEST_LV_0100.tsv
Some more updates:

QUEST_LV_0100_20150317_000148 - Corrected grammatical error. Added the word "old" for consistency due to an older mention of it.

QUEST_LV_0100_20150317_000153 - The word "canyon" is not capitalized in earlier mentions and seems to be a common name.

QUEST_LV_0100_20150317_000184 - Punctuation error. A comma is necessary in this question.

QUEST_LV_0100_20150317_000197 - Another comma is needed here.

QUEST_LV_0100_20150317_000205 - I believe a comma should not be there as there's no need to separate the sentence in two.

QUEST_LV_0100_20150317_000210 - Ellipses are punctuation marks consisting of three dots.

QUEST_LV_0100_20150317_000211 - Same as above.

QUEST_LV_0100_20150317_000213 - Grammatical error.

QUEST_LV_0100_20150317_000231 - Added an extra dot to the ellipsis.

QUEST_LV_0100_20150317_000239 - Corrected ellipses.

QUEST_LV_0100_20150317_000241 - Why is "old grains"  capitalized? There's another reference in this quest line as a common description and it is not capitalized.

QUEST_LV_0100_20150317_000269 - A comma is unnecessary there.

QUEST_LV_0100_20150317_000275 - Replaced "be" with "the".

QUEST_LV_0100_20150317_000278 - What is this {-scp SCR_DIALOG_NPC_ANIM(? Is this a command and thus correct? Clarification for future reference would be nice.