Update QUEST_LV_0100.tsv
So more updates as I continue to read:

QUEST_LV_0100_20150317_000311 - Comma was needed.

QUEST_LV_0100_20150317_000313 - Same as above.

QUEST_LV_0100_20150317_000318 - Same as above.

QUEST_LV_0100_20150317_000321 - Same as above and changed the word "out" as it was in wrong context.

QUEST_LV_0100_20150317_000331 - No need for a comma there.

QUEST_LV_0100_20150317_000333 - Same as above.

QUEST_LV_0100_20150317_000338 - Put a comma.

QUEST_LV_0100_20150317_000339 - Capitalized the name of a monster as its been done before. It's the Vubbe Figher, not just any Vubbe fighter.

QUEST_LV_0100_20150317_000340 - Same as above.

QUEST_LV_0100_20150317_000343 - No comma needed.

QUEST_LV_0100_20150317_000346 - Comma added.

QUEST_LV_0100_20150317_000355 - Same as above.

QUEST_LV_0100_20150317_000358 - The officer is granting you access, not granting something else to your access. It would sound better to make use of "your" if it was something like "I will allow your access." or "I will grant more privileges to your access." However, in this case, I think "you" is more appropriate.

QUEST_LV_0100_20150317_000358 - Comma added.

QUEST_LV_0100_20150317_000361 - Added a comma and replaced the word "build". This word is more associated with architecture. If talking about putting people or things together then to "form" or "create" a group sounds more fitting to the situation.

QUEST_LV_0100_20150317_000367 - Capitalized for consistency.

QUEST_LV_0100_20150317_000368 - This is my suggestion because as it is right now it makes no sense without the hyphen. I also replaced "it" with "them" because he referred to stones (as in many). If he is referring to the Tablet of Revelation that you acquire in the area then it should be clarified in the word "stones".

QUEST_LV_0100_20150317_000375 - Comma added.
QUEST_LV_0100_20150317_000376 - Comma added.

QUEST_LV_0100_20150317_000384 - Changed it to the village name. It would not be correct to capitalize the common reference to the mining village.

QUEST_LV_0100_20150317_000390 - Village name is missing an apostrophe.

QUEST_LV_0100_20150317_000394 - I believe "further down" is a proper way to say it as it is not something underground in this case.

QUEST_LV_0100_20150317_000400 - Corrected wrong comma placement.