Hello, instead of sitting around and waiting for this awesome game to be released, I would like to help out with the translations.
I just did a few lines to see if I'm compatible with the project. If not,
please let me know the things I have to change to help this process
smoother.

20150317_003413 & 003472

I tried to fill in the gaps, then read the Guideline. I jumped to an empty
chunk on the bottom

20140707_005530 - 20140707_005532

I have placed an "x" before the lines.