Forest of Prayer Altar Typo
I know this is a one line edit, but I've been quite busy with testing in-game content and working, so I haven't had the time to sit down and do a batch (hopefully I'll get something done when I have my day off).  However, this is a really blatant typo that I noticed while screwing around with the quest related to this line, so it's pretty isolated to begin with.

Once again, sorry about this, but it's a really ugly issue that is a quick fix.