Update QUEST.tsv [Lines 1651-1800]
Lines 1651-1800 reviewed. Lines that disappeared after the merge to CBT re-added. Open to feedback.

Again, changed "contaminating gap" to "corruption breach." If anyone can think of a better keyword or think the original is fine then I'm okay with changing it to that or back to the original.

I'd like some input before this is merged.