Update ITEM.tsv
Changed to fix the weird spaces and auto changes added to lines above what I actually edited.  Also fixed a space on the very first entry?  Not sure if that was supposed to be there or not, but I deleted it.  Please refer to my first, main memo for my questions and explanations.