Update QUEST_JOBSTEP.tsv
Lines  3072 - 3138

Line 3091:  Just to be clear, these 3 Masters are "Amjinases?"  I'm assuming Amjinases is plural and just the Necro Master would be an Amjinas.

Line 3101: The second sentence is really odd, but I can't exactly tell what the context is (aside from the orb and exploration thing).  What is dangerous?  The mines?  The orb?  I changed the wording a bit to make it more fluid, but it's still kind of confusing.

Line 3128 (Line 3087 too): Not sure what "31_1" is?  I took out the second "the first Fletcher Master," as it's a bit redundant.

Line 3125: I'm like 98% sure the Scout Master in that map is female.