Small update - just a few questions and changes.
Going up still.  Stopped at line 6026 for this commit.  Will work on more stuff later tonight.

Lines 6054-56: What is this?  Just curious. Some sort of scaling thing?

Line 6065: It's a multiple level attribute, so I think it should include per level

Line 6069:  Should "continuously" be removed an there be something like"it will apply the [Attacked Weakened] status effect with each hit?"  Continuously makes it seem like a tick thing, which if it is, please correct me.

Line 6088: This wording is really weird...should it be "Increases critical resistance based on a percentage of the caster's SPR when using [Rune of Protection] by 20% per attribute level?"