QUEST_LV_0200_20150317_002844
QUEST_LV_0200_20150317_002844	I have collected all the materials. Dievdirbys master Let's try to repair is carried out with the material to the image as precipitation.

QUEST_LV_0200_20150317_002844	You have all the materials! Go visit the Dievdirbys Master to carry out repairs on the statue.

This is a mess. Not a Korean speaker but I've done my best with it.